Conjugacy Classes

For some groups (particularly matrix groups) the concept of conjugacy classes is important.

The conjugacy class of an element $g \in \mathscr{G}$ is defined as

\[Cl(g) = \{h \cdot g \cdot h^{-1} | h \in \mathscr{G}\}\]

where $h^{-1}$ is the inverse element inv(h) to h with $h\cdot h^{-1} = e$ where $e$ is the identity element of the group.

This package provides some functions to handle conjugacy classes.

Calculating all conjugacy classes

The conjugacy classes can be calculated either using Julia's internal inv() function, or by providing a function as the inverse parameter.

ClosedGroupFunctions.calculate_conjugacy_classes — Method

calculate_conjugacy_classes(group::Bijection{String, T}; inverse=inv, prnt=true) -> conjugacy_classes::Vector{Set{String}}

Calculate all conjugacy classes {Cl(g) = {h * g * h^{-1} | h ∈ 𝒢}} for a given finite, closed group, given the inverse function for elements of type T in the group group. The inverse function defaults to the Julia inv function. prnt enables/disables output of the class-sizes and remaining elements while computing the classes.

Info

Every element occurs only in one conjugacy classes. A disjoint union ($\dot\cup$) of all conjugacy classes of $\mathscr{G}$ yields again $\mathscr{G}$.

Investigating classes

Conjugacy classes or any other sub-list of labelled elements of a labelled group can be investigated further by applying functions, here called an invariant because of the connection with conjugacy classes and functions, that are constant for a given class (but every function that expects only a group element of type T can be used).

Is function invariant?

ClosedGroupFunctions.is_invariant — Method

is_invariant(class::Set{String}, group::Bijection{String, T}, invariant) -> (is_invariant::Bool, values::Vector)

Verifies that the given function invariant is an invariant for a given (conjugacy) class. The class is given by labels that match the corresponding elements in the Bijection group.

Applying a function to the class(es)

To every element of one class

ClosedGroupFunctions.apply_invariant_to_all_in_class — Method

apply_invariant_to_all_in_class(class::Set{String}, group::Bijection{String, T}, invariant; prnt=true) -> values::Vector

Calculates the values of invariant for every element of a given class. The class is given by labels that match the corresponding elements in the Bijection group.

To the first element of one class

ClosedGroupFunctions.apply_invariant_to_first_in_class — Method

apply_invariant_to_first_in_class(class::Set{String}, group::Bijection{String, T}, invariant) -> value

Calculates the value of invariant for the first element of a given class. The class is given by labels that match the corresponding elements in the Bijection group.

To the first element of every given class

ClosedGroupFunctions.apply_invariant_to_first_in_all_classes — Method

apply_invariant_to_first_in_all_classes(conjugacy_classes::Vector{Set{String}}, group::Bijection{String, T}, invariant; prnt=true) -> values::Vector

Calculates the values of invariant for the first element of all given classes. The conjugacy_classes are given by Sets of labels that match the corresponding elements in the Bijection group.