Generating and Labelling

Assuming you already have a list of generators of a finite group, the group under closure can be calculated with the following functions.

Basic generator

The basic, very efficient group generator takes an array of generators of type T and yields a tuple with a Set of all group elements of type T and the number of multiplications that were needed to calculate the whole group.

ClosedGroupFunctions.group_generator_basic — Method

group_generator_basic(generators::Array{T}; prnt=false) -> Set(group), num_all_mult

Calculate all other elements of this closed group by multiplying all generators and storing new elements and then multiplying the generators again with the new elements. Very efficient with $O(\Omega \cdot N)$ multiplications. It is the most efficient algorithm for a given set of generators. Only the use of smaller sets of generators can reduce the number of multiplications further.

The prnt variable can be used to specify if the programm should output the progressbar. This can slow things down.

Labelling

Labelling of the generators

A given array of generators of type T can be converted to a labelled Bijection using

ClosedGroupFunctions.label_generators — Method

label_generators(generators::Array{T}; alphabet::Vector{Char}) -> labelled_generators::Bijection{String, T}

Labels the given generators using the given alphabet, a vector of characters and returns the labelled generators. The alphabet can be bigger than the number of generators, only the first Ω = length(generators) characters are used.

Info

There the default_alphabet() returns an alphabet with 103 possible characters that can be used for labelling (For bigger generator-lists, please use your own character list). The default alphabet consists of 'a':'z', 'A':'Z', '𝒜':'𝒵' and 'α':'ω'.

Labelling of the group

There are two functions, that can be used to label the whole group.

The 'simple' generator is fast and stores the first occuring label of every element. It stops if every element is calculated - for this to work, the final number of elements (from the result of the basic generator) has to be provided.

ClosedGroupFunctions.labelled_group_generator_simple — Method

labelled_group_generator_simple(labelled_generators::Bijection{String, T}, num_max_elements::Int64; prnt = false) -> labelled_group::Bijection{String, T}

Fast Method: Generates all group elements and labels them with the first occuring label. It stops after generating num_max_elements elements. The resulting label may not be the shortest possible label. The generators can be given to the function either as a labelled Bijection{String, T} or as an unlabelled Array{T} that will automatically labelled.

The 'shortest' method works analogous to the basic generator - it calculates every element and every possible multiplication and stores only the shortest occuring label of every element. This method can be quite slow because the difference in the storage: while the basic generator uses a Set{T} that automatically only keeps one copy of every element, the labelled generators use a Bijection{String, T} that is a bit slower. Aside from that, every time an element is calculated the function needs to check, if the new label is shorter than the stored one and switch the entries if so.

ClosedGroupFunctions.labelled_group_generator_shortest — Method

labelled_group_generator_shortest(labelled_generators::Bijection{String, T}; prnt = false)::Tuple{Bijection{String, T}, Int64} -> labelled_group::Bijection{String, T}, num_all_mult::Int64

For the general method and description of prnt see group_generator_basic(). Slow Method: This method generates all other elements of the closed group and labels them using the shortest possible label. As Bijections are slower than Sets and replacing the longer-labelled Pair with the shorter-labelled Pair is time consuming, this method can be really slow compared to group_generator_basic().